requirements.txt
flask
flask-login
google-generativeai
pandas
